+++ src/app/windowgeometry.cpp
@@ -5,26 +5,23 @@ namespace Nickvision::App
     WindowGeometry::WindowGeometry()
         : m_width{ 800 },
         m_height{ 600 },
-        m_isMaximized{ false }
+        m_isMaximized{ false },
+        m_x{ 10 },
+        m_y{ 10 }
     {
-#ifdef _WIN32
-        m_x = 10;
-        m_y = 10;
-#endif
+
     }
 
     WindowGeometry::WindowGeometry(long width, long height, bool isMaximized)
         : m_width{ width },
         m_height{ height },
-        m_isMaximized{ isMaximized }
+        m_isMaximized{ isMaximized },
+        m_x{ 10 },
+        m_y{ 10 }
     {
-#ifdef _WIN32
-        m_x = 10;
-        m_y = 10;
-#endif
+
     }
 
-#ifdef _WIN32
     WindowGeometry::WindowGeometry(long width, long height, bool isMaximized, long x, long y)
         : m_width{ width },
         m_height{ height },
@@ -35,27 +32,14 @@ namespace Nickvision::App
 
     }
 
-    WindowGeometry::WindowGeometry(HWND hwnd)
-    {
-        WINDOWPLACEMENT placement;
-        GetWindowPlacement(hwnd, &placement);
-        m_width = placement.rcNormalPosition.right - placement.rcNormalPosition.left;
-        m_height = placement.rcNormalPosition.bottom - placement.rcNormalPosition.top;
-        m_isMaximized = placement.showCmd == SW_SHOWMAXIMIZED;
-        m_x = placement.rcNormalPosition.left;
-        m_y = placement.rcNormalPosition.top;
-    }
-#endif
-
     WindowGeometry::WindowGeometry(boost::json::object json)
-        : m_width{ json["Width"].is_int64() ? static_cast<long>(json["Width"].as_int64()) : 0 },
-        m_height{ json["Height"].is_int64() ? static_cast<long>(json["Height"].as_int64()) : 0 },
-        m_isMaximized{ json["IsMaximized"].is_bool() ? json["IsMaximized"].as_bool() : false }
+        : m_width{ json["Width"].is_int64() ? static_cast<long>(json["Width"].as_int64()) : 800 },
+        m_height{ json["Height"].is_int64() ? static_cast<long>(json["Height"].as_int64()) : 600 },
+        m_isMaximized{ json["IsMaximized"].is_bool() ? json["IsMaximized"].as_bool() : false },
+        m_x{ json["X"].is_int64() ? static_cast<long>(json["X"].as_int64()) : 10 },
+        m_y{ json["Y"].is_int64() ? static_cast<long>(json["Y"].as_int64()) : 10 }
     {
-#ifdef _WIN32
-        m_x = json["X"].is_int64() ? static_cast<long>(json["X"].as_int64()) : 0;
-        m_y = json["Y"].is_int64() ? static_cast<long>(json["Y"].as_int64()) : 0;
-#endif
+
     }
 
     long WindowGeometry::getWidth() const
@@ -88,19 +72,25 @@ namespace Nickvision::App
         m_isMaximized = isMaximized;
     }
 
-#ifdef _WIN32
-    bool WindowGeometry::apply(HWND hwnd) const
+    long WindowGeometry::getX() const
     {
-        WINDOWPLACEMENT placement;
-        GetWindowPlacement(hwnd, &placement);
-        placement.rcNormalPosition.left = m_x;
-        placement.rcNormalPosition.top = m_y;
-        placement.rcNormalPosition.right = placement.rcNormalPosition.left + m_width;
-        placement.rcNormalPosition.bottom = placement.rcNormalPosition.top + m_height;
-        placement.showCmd = m_isMaximized ? SW_SHOWMAXIMIZED : SW_SHOWNORMAL;
-        return SetWindowPlacement(hwnd, &placement);
+        return m_x;
     }
-#endif
+
+    void WindowGeometry::setX(long x)
+    {
+        m_x = x;
+    }
+
+    long WindowGeometry::getY() const
+    {
+        return m_y;
+    }
+
+    void WindowGeometry::setY(long y)
+    {
+        m_y = y;
+    }
 
     boost::json::object WindowGeometry::toJson() const
     {
@@ -108,10 +98,8 @@ namespace Nickvision::App
         json["Width"] = m_width;
         json["Height"] = m_height;
         json["IsMaximized"] = m_isMaximized;
-#ifdef _WIN32
         json["X"] = m_x;
         json["Y"] = m_y;
-#endif
         return json;
     }
 }